requirements.txt
streamlit
openai
llama-index
nltk
python-dotenv
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
chromadb
faiss-cpu
langchain_google_genai
langchain-community